requirements.txt
Flask==0.10.1
requests==2.6.2
optimizely-sdk==2.1.1
